moved file location to components folder
